REST CRUD 102 full CRUD

+++ 04-spring-boot-rest-crud/02-spring-boot-rest-crud-employee/cruddemo/src/main/java/com/dev/cruddemo/rest/EmployeeRestController.java
@@ -91,6 +91,10 @@ public class EmployeeRestController {
     @DeleteMapping("/employees/{id}")
     public void deleteEmployee(@PathVariable int id) {
 
+        Employee tempEmployee = service.findById(id);
+        if (tempEmployee == null)
+            throw new RuntimeException("Employee with id: " + id + " not exist.");
+
         service.deleteEmployeeById(id);
     }
 }
